+++ src/component/atom/link/TrialLink.tsx
@@ -1,5 +1,5 @@
 import { Link } from '@chakra-ui/react';
-import { useAppDispatch } from '@/redux/store';
+import { useAppDispatch } from '@/redux/hooks';
   
 export default function TrialLink({ ...props }) {
