+++ src/redux/store.ts
@@ -1,10 +1,7 @@
-import { AnyAction, ThunkAction, ThunkDispatch, configureStore } from '@reduxjs/toolkit';
-
-import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
+import { AnyAction, ThunkAction, configureStore } from '@reduxjs/toolkit';
 
 import web3Reducer from './features/web3';
 import appReducer from './features/app';
-
 
 export const store = configureStore({
   reducer: {
@@ -27,9 +24,4 @@ export const store = configureStore({
 // Infer the `RootState` and `AppDispatch` types from the store itself
 export type RootState = ReturnType<typeof store.getState>;
 export type AppDispatch = typeof store.dispatch;
-
-// Use throughout your app instead of plain `useDispatch` and `useSelector`
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
-export const useAppDispatch = useDispatch<ThunkDispatch<RootState, unknown, AnyAction>>;
-
 export type AppThunk<ReturnType = void> = ThunkAction<ReturnType, RootState, unknown, AnyAction>
